Update Badger
Remove check in badger.award() for if the player has the badge already, as this should be implemented by the developer as an if statement before making the call to award. This is being done to make the success bool returned, will always be a BadgeService error if its false.